Payment

+++ restaurant/src/main/java/cln/swiggy/restaurant/serviceImpl/MenuServiceImpl.java
@@ -1,12 +1,14 @@
 package cln.swiggy.restaurant.serviceImpl;
 
 import cln.swiggy.restaurant.exception.ResourceNotFoundException;
+import cln.swiggy.restaurant.model.Category;
 import cln.swiggy.restaurant.model.Menu;
 import cln.swiggy.restaurant.model.MenuImage;
 import cln.swiggy.restaurant.model.Restaurant;
 import cln.swiggy.restaurant.model.request.MenuRequest;
 import cln.swiggy.restaurant.model.response.CommonResponse;
 import cln.swiggy.restaurant.model.response.MenuResponse;
+import cln.swiggy.restaurant.repository.CategoryRepository;
 import cln.swiggy.restaurant.repository.MenuImagesRepository;
 import cln.swiggy.restaurant.repository.MenuRepository;
 import cln.swiggy.restaurant.repository.RestaurantRepository;
@@ -15,6 +17,7 @@ import cln.swiggy.restaurant.serviceImpl.otherImple.ImageUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -27,6 +30,7 @@ public class MenuServiceImpl implements MenuService {
     @Autowired  MenuRepository menuRepository;
     @Autowired  RestaurantRepository restaurantRepository;
     @Autowired  MenuImagesRepository menuImageRepository;
+    @Autowired  CategoryRepository categoryRepository;
     @Autowired  ImageUtils imageUtils;
 
     @Override
@@ -34,8 +38,18 @@ public class MenuServiceImpl implements MenuService {
             Restaurant restaurant = restaurantRepository.findById(restaurantId)
                     .orElseThrow(() -> new ResourceNotFoundException("Restaurant not found with id: " + restaurantId));
 
-            Menu menu = new Menu();
+            Category category = categoryRepository.findById(request.getCategoryId())
+                    .orElseThrow(() -> new ResourceNotFoundException("Category not found with id: " + request.getCategoryId()));
+
+            if (menuRepository.existsByNameAndRestaurantAndPrice(request.getName(), restaurant, request.getPrice())) {
+                return ResponseEntity.badRequest()
+                        .body(new CommonResponse(400, "Menu item with same name and price already exists in this restaurant", false));
+            }
+
+
+        Menu menu = new Menu();
             menu.setRestaurant(restaurant);
+            menu.setCategory(category);
             menu.setName(request.getName());
             menu.setDescription(request.getDescription());
             menu.setPrice(request.getPrice());
@@ -46,18 +60,16 @@ public class MenuServiceImpl implements MenuService {
 
             Menu savedMenu = menuRepository.save(menu);
 
-            if (request.getImages() != null && !request.getImages().isEmpty()) {
-                List<String> imagePaths = imageUtils.saveImages(request.getImages(), "menu-items");
-                imagePaths.forEach(path -> {
-                    MenuImage image = new MenuImage();
-                    image.setMenu(savedMenu);
-                    image.setImage(path);
-                    menuImageRepository.save(image);
-                });
+        if (request.getImages() != null && !request.getImages().isEmpty()) {
+            for (MultipartFile file : request.getImages()) {
+//              String key = storageService.uploadFile(file);
+                MenuImage image = new MenuImage();
+                image.setMenu(savedMenu);
+                image.setImage("menu.jpg");
+                menuImageRepository.save(image);
             }
-
-            return ResponseEntity.ok(new CommonResponse(201, "Menu item created successfully",
-                    MenuResponse.convertToResponse(menuRepository.save(savedMenu))));
+        }
+            return ResponseEntity.ok(new CommonResponse(201, "Menu item created successfully",true));
     }
 
     @Override
@@ -68,10 +80,9 @@ public class MenuServiceImpl implements MenuService {
                 return ResponseEntity.ok(new CommonResponse(200, "No menu items found", new ArrayList<>()));
             }
 
-            List<MenuResponse> responses = menuItems.stream()
-                    .map(MenuResponse::convertToResponse)
-                    .collect(Collectors.toList());
-
+        List<MenuResponse> responses = menuItems.stream()
+                .map(menu -> MenuResponse.convertToResponse(menu, menuImageRepository))
+                .collect(Collectors.toList());
             return ResponseEntity.ok(new CommonResponse(200, "Menu items retrieved successfully", responses));
     }
 
@@ -81,7 +92,7 @@ public class MenuServiceImpl implements MenuService {
                     .orElseThrow(() -> new ResourceNotFoundException("Menu item not found with id: " + menuItemId));
 
             return ResponseEntity.ok(new CommonResponse(200, "Menu item retrieved successfully",
-                    MenuResponse.convertToResponse(menuItem)));
+                    MenuResponse.convertToResponse(menuItem,menuImageRepository)));
     }
 
     @Override
@@ -89,14 +100,23 @@ public class MenuServiceImpl implements MenuService {
             Menu menu = menuRepository.findById(menuItemId)
                     .orElseThrow(() -> new ResourceNotFoundException("Menu item not found with id: " + menuItemId));
 
+            Category category = categoryRepository.findById(request.getCategoryId())
+                .orElseThrow(() -> new ResourceNotFoundException("Category not found with id: " + request.getCategoryId()));
+
+            if (menuRepository.existsByNameAndRestaurantAndPrice(request.getName(), menu.getRestaurant(), request.getPrice())) {
+                return ResponseEntity.badRequest()
+                        .body(new CommonResponse(400, "Menu item with same name and price already exists in this restaurant", false));
+            }
+
             menu.setName(request.getName());
+            menu.setCategory(category);
             menu.setDescription(request.getDescription());
             menu.setPrice(request.getPrice());
             menu.setDiscount(request.getDiscount());
             menu.setUpdatedAt(LocalDateTime.now());
 
             if (request.getImages() != null && !request.getImages().isEmpty()) {
-                List<MenuImage> existingImages = menuImageRepository.findByMenuId(menuItemId);
+                List<MenuImage> existingImages = menuImageRepository.findByMenu(menu);
                 existingImages.forEach(image -> {
                     imageUtils.deleteImage(image.getImage(), "menu-items");
                     menuImageRepository.delete(image);
@@ -111,8 +131,7 @@ public class MenuServiceImpl implements MenuService {
                 });
             }
 
-            return ResponseEntity.ok(new CommonResponse(200, "Menu item updated successfully",
-                    MenuResponse.convertToResponse(menuRepository.save(menu))));
+            return ResponseEntity.ok(new CommonResponse(200, "Menu item updated successfully",true));
     }
 
     @Override
@@ -120,7 +139,7 @@ public class MenuServiceImpl implements MenuService {
             Menu menu = menuRepository.findById(menuItemId)
                     .orElseThrow(() -> new ResourceNotFoundException("Menu item not found with id: " + menuItemId));
 
-            List<MenuImage> images = menuImageRepository.findByMenuId(menuItemId);
+            List<MenuImage> images = menuImageRepository.findByMenu(menu);
             images.forEach(image -> {
                 imageUtils.deleteImage(image.getImage(), "menu-items");
                 menuImageRepository.delete(image);
